Refactor evaluations to fit into i16 (#453)
Bench: 2082244

+++ src/types.h
@@ -11,7 +11,7 @@
 #endif
 
 // General constants and types
-using Eval = int32_t;
+using Eval = int16_t;
 using Square = uint8_t;
 using Bitboard = uint64_t;
 using Depth = int16_t;
